Move lib directory up one
